Create getStaker Info func
