+++ services/fadostaking/index.ts
@@ -13,8 +13,18 @@ import { FADO_STAKE_SMC, FADO_TOKEN_SMC } from './config';
 const kardiaClient = new KardiaClient({endpoint: RPC_ENDPOINT});
 const kardiaContract = kardiaClient.contract;
 
-// Base on ABI defined, note that function with constant = "false" > invoke.send ,
-// else invoke.call
+export const getStakerInfo = async (walletAddress: string) => {
+  kardiaContract.updateAbi(FADO_STAKING_ABI);
+  const response = await kardiaContract.invokeContract('stakerInfo', [walletAddress]).call(FADO_STAKE_SMC);
+
+  console.log({response});
+}
+
+/**
+ * @param stakeAmount 
+ * @param wallet 
+ * @returns tx : Return transaction address.
+ */
 export const stakeFadoToken = async ( stakeAmount: number , wallet: Wallet) => {
   const convertAMount = cellValue(stakeAmount);
   console.log({convertAMount});
@@ -24,6 +34,8 @@ export const stakeFadoToken = async ( stakeAmount: number , wallet: Wallet) => {
     const approveInvocation = await kardiaContract.invokeContract('approve', [FADO_STAKE_SMC , convertAMount]);
     // Test fail thu switch lại
     const estimatedGas = await approveInvocation.estimateGas(approveInvocation.getDefaultTxPayload)
+    console.log({estimatedGas});
+    
     const approveStatus = await approveInvocation.send(wallet.privateKey!, FADO_TOKEN_SMC,{
       from: wallet.address,
       gas: DEFAULT_GAS_LIMIT,
@@ -35,6 +47,8 @@ export const stakeFadoToken = async ( stakeAmount: number , wallet: Wallet) => {
     kardiaContract.updateAbi(FADO_STAKING_ABI);
     const stakeInvocation = await kardiaContract.invokeContract('stake', [convertAMount])
     const stakeEstimateGas = await stakeInvocation.estimateGas(stakeInvocation.getDefaultTxPayload)
+    console.log({stakeEstimateGas});
+
     const txtAddress = await stakeInvocation.send(wallet.privateKey!, FADO_STAKE_SMC, {
       from: wallet.address,
       gas: DEFAULT_GAS_LIMIT,
@@ -49,9 +63,20 @@ export const stakeFadoToken = async ( stakeAmount: number , wallet: Wallet) => {
       }      
 }
 
+
+/**
+ * 
+ * @param wallet,@param smcAddr
+ * @return numberAllowance: 
+ * This function get back the approved amount that was call in 'approve' 
+ */
 const allowance = async (wallet: Wallet, smcAddr: string) => {
+  console.log("HELLO");
+  
   kardiaContract.updateAbi(FADO_TOKEN_ABI);
   const numberAllowance = await kardiaContract.invokeContract('allowance', [wallet.address, smcAddr]).call(FADO_TOKEN_SMC);
 
   console.log({numberAllowance});
 }
+
+
